requirements.txt
spacy
spacy-transformers
torch
pandas
datasets
python-dotenv
tqdm

openai
